port settings to static viz
